+++ sysbridge/src/main/java/io/github/sds100/keymapper/sysbridge/utils/IContentProviderUtils.kt
@@ -2,6 +2,8 @@ package io.github.sds100.keymapper.sysbridge.utils
 
 import android.content.AttributionSource
 import android.content.IContentProvider
+import android.database.Cursor
+import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 
@@ -37,5 +39,29 @@ internal object IContentProviderUtils {
         }
 
         return result
+    }
+
+    @Throws(android.os.RemoteException::class)
+    fun queryCompat(
+        provider: IContentProvider,
+        callingPkg: String?,
+        url: Uri,
+        projection: Array<String>?,
+        queryArgs: Bundle?,
+    ): Cursor? {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            val uid = android.system.Os.getuid()
+            provider.query(
+                AttributionSource.Builder(uid).setPackageName(callingPkg).build(),
+                url,
+                projection,
+                queryArgs,
+                null,
+            )
+        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            provider.query(callingPkg, null as String?, url, projection, queryArgs, null)
+        } else {
+            provider.query(callingPkg, url, projection, queryArgs, null)
+        }
     }
 }